+++ ffi-client/examples/triangle_headless.rs
@@ -60,7 +60,7 @@ fn main() -> goldy_ffi_client::Result<()> {
     )?;
 
     let mut scheme = Scheme::new(&ctx)?;
-    let rt = scheme.lease_render_target(WIDTH, HEIGHT, TextureFormat::Rgba8Unorm, None::<DepthFormat>)?;
+    let rt = ctx.lease_render_target(WIDTH, HEIGHT, TextureFormat::Rgba8Unorm, None::<DepthFormat>)?;
     {
         let mut pass = scheme.render_pass("triangle", &rt, TargetLoad::Clear(Color::BLACK));
         pass.with_buffer(&vertex_buffer, NodeAccess::Read);